+++ demisto_sdk/commands/common/content/objects/pack_objects/pack.py
@@ -30,8 +30,8 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
     Contributors, Dashboard, DocFile, GenericDefinition, GenericField,
     GenericModule, GenericType, IncidentField, IncidentType, IndicatorField,
     IndicatorType, Integration, LayoutObject, OldClassifier, PackIgnore,
-    PackMetaData, Playbook, Readme, ReleaseNote, Report, Script, SecretIgnore,
-    Widget)
+    PackMetaData, Playbook, Readme, ReleaseNote, ReleaseNoteConfig, Report,
+    Script, SecretIgnore, Widget)
 from demisto_sdk.commands.common.content.objects_factory import \
     path_to_pack_object
 from demisto_sdk.commands.test_content import tools
@@ -158,6 +158,11 @@ class Pack:
     def release_notes(self) -> Iterator[ReleaseNote]:
         return self._content_files_list_generator_factory(dir_name=RELEASE_NOTES_DIR,
                                                           suffix="md")
+
+    @property
+    def release_notes_config(self) -> Iterator[ReleaseNoteConfig]:
+        return self._content_files_list_generator_factory(dir_name=RELEASE_NOTES_DIR,
+                                                          suffix="json")
 
     @property
     def generic_definitions(self) -> Iterator[GenericDefinition]: